wasmer: move Memory definition to the wasmer-specific file
All other VMs have their memories defined within their `_runner` module.
No reason why the old Wasmer0 codebase needs special treatment here.

(It was also bothering me when I would see a generic `memory.rs` and
upon opening it get the wasmer file, rather the file that describes the
`MemoryLike` trait...
